fix(lints): clippy -D warnings clean across every CI invocation (#39)
With the superwork compile fix landed, CI clippy lints the real tree for
the first time in weeks and surfaced a backlog. All five CI clippy lines
are now locally green (workspace default + no-default, zenfilters,
zenlayout, zencodecs no-default):

zencodecs:
- crate-level cfg_attr: with every codec feature off (zenpipe builds
  zencodecs default-features=false until nodes-* enable codecs) the
  dispatch surface and feature-gated test bodies are dead by
  construction — relax dead-code lints for that combination only
- [[test]] required-features for metadata_conformance / corpus /
  stop_latency (codec-driven binaries, empty shells without codecs)

zenpipe:
- is_multiple_of modernization (composite, windowed_filter)
- useless format! -> String::from (effects)
- unnecessary usize cast (sidecar test)
- allow(deprecated) + migration notes for the ColorManagement surface
  (PluggableCms migration is a tracked API change, re-export is public)
- allow(dead_code) + notes: solid_background_from_canvas_color (canvas
  wiring pending), EstimateInfo::add_streaming (audit-M5 wiring pending)
- allow(too_many_arguments): transcode_with_stop_and_limits (public
  audit-H8 entry point; regrouping params is an API break)

zenfilters:
- ai_corpus_flatten example: doc-list paragraph break +
  allow(collapsible_if) (let-chains would break MSRV)

Verified: fmt --check, workspace tests default+no-default, zencodecs
no-default + wide gain-map set — all green.

+++ src/sidecar.rs
@@ -334,7 +334,7 @@ mod tests {
 
         // Read per-pixel values regardless of what layout the pipeline
         // emitted (channel 0 is the gain value in both Gray and RGB).
-        let ch = mat.format().channels() as usize;
+        let ch = mat.format().channels();
         let values: alloc::vec::Vec<u8> = mat
             .data()
             .chunks_exact(mat.format().aligned_stride(mat.width()))
